refactor: splitting RealmCache file in many files, one file for each symbol
