+++ Kinvey/Kinvey/RealmCache.swift
@@ -16,7 +16,7 @@ fileprivate let typeIntValue = IntValue.self.className()
 fileprivate let typeFloatValue = FloatValue.self.className()
 fileprivate let typeDoubleValue = DoubleValue.self.className()
 fileprivate let typeBoolValue = BoolValue.self.className()
-fileprivate let typesNeedsTranslation = [
+internal let typesNeedsTranslation = [
     typeStringValue,
     typeIntValue,
     typeFloatValue,
@@ -37,17 +37,7 @@ fileprivate let typesNeedsTranslation = [
     }
 #endif
 
-public enum CollectionChange<CollectionType> {
-    
-    case initial(CollectionType)
-    
-    case update(CollectionType, deletions: [Int], insertions: [Int], modifications: [Int])
-    
-    case error(Swift.Error)
-    
-}
-
-internal class RealmCache<T: Persistable>: Cache<T>, CacheType where T: NSObject {
+internal class RealmCache<T: Persistable>: Cache<T>, CacheType, RealmCascadeDeletable where T: NSObject {
     
     typealias `Type` = T
     
@@ -691,7 +681,7 @@ internal class RealmCache<T: Persistable>: Cache<T>, CacheType where T: NSObject
         self.newRealm.cancelWrite()
     }
     
-    private func write(_ block: @escaping (Realm) throws -> Void) throws {
+    internal func write(_ block: @escaping (Realm) throws -> Void) throws {
         if newRealm.isInWriteTransaction {
             var _error: Swift.Error? = nil
             do {
@@ -713,207 +703,6 @@ internal class RealmCache<T: Persistable>: Cache<T>, CacheType where T: NSObject
 }
 
 extension RealmSwift.NotificationToken: NotificationToken {
-}
-
-extension RealmCache: DynamicCacheType {
-    
-    private func countReferences(realm: Realm, objectClassName: String, nestedObject: Object, threshold: Int? = nil) -> Int {
-        var count = 0
-        for schema in realm.schema.objectSchema {
-            guard schema.className != objectClassName else {
-                continue
-            }
-            
-            for property in schema.properties {
-                guard property.objectClassName == objectClassName else {
-                    continue
-                }
-                
-                if property.isArray {
-                    count += realm.dynamicObjects(schema.className).filter("ANY \(property.name) == %@", nestedObject).count
-                } else {
-                    count += realm.dynamicObjects(schema.className).filter("\(property.name) == %@", nestedObject).count
-                }
-                if let threshold = threshold, count > threshold {
-                    return count
-                }
-            }
-        }
-        return count
-    }
-    
-    internal func cascadeDelete(realm: Realm, entityType: String, entity: Object, deleteItself: Bool = true) {
-        if let cascadeDeletable = entity as? CascadeDeletable {
-            deleteAclAndMetadata(realm: realm, object: entity)
-            try! cascadeDeletable.cascadeDelete(executor: RealmCascadeDeleteExecutor(realm: realm))
-        } else if let schema = realm.schema[entityType] {
-            signpost(.begin, log: osLog, name: "Cascade Delete", "%@", entityType)
-            defer {
-                signpost(.end, log: osLog, name: "Cascade Delete", "%@", entityType)
-            }
-            schema.properties.forEachAutoreleasepool { property in
-                switch property.type {
-                case .object:
-                    if property.isArray,
-                        let primaryKeyProperty = schema.primaryKeyProperty,
-                        let entityId = entity[primaryKeyProperty.name],
-                        let dynamicObject = realm.dynamicObject(ofType: entityType, forPrimaryKey: entityId),
-                        let objectClassName = property.objectClassName,
-                        let nestedArray = dynamicObject[property.name] as? List<DynamicObject>
-                    {
-                        cascadeDelete(
-                            realm: realm,
-                            entityType: objectClassName,
-                            entities: nestedArray
-                        )
-                    } else if let objectClassName = property.objectClassName,
-                        let nestedObject = entity[property.name] as? Object,
-                        countReferences(realm: realm, objectClassName: objectClassName, nestedObject: nestedObject, threshold: 1) == 1
-                    {
-                        cascadeDelete(
-                            realm: realm,
-                            entityType: objectClassName,
-                            entity: nestedObject
-                        )
-                    }
-                default:
-                    break
-                }
-            }
-        }
-        if deleteItself {
-            realm.delete(entity)
-        }
-    }
-    
-    private func cascadeDelete(realm: Realm, entityType: String, entities: List<DynamicObject>) {
-        for entity in entities {
-            guard countReferences(realm: realm, objectClassName: entityType, nestedObject: entity, threshold: 1) == 1 else {
-                continue
-            }
-            cascadeDelete(
-                realm: realm,
-                entityType: entityType,
-                entity: entity
-            )
-        }
-    }
-    
-    private func cascadeDelete(realm: Realm, entityType: String, entity: JsonDictionary, propertyMapping: PropertyMap) {
-        propertyMapping.forEachAutoreleasepool { translatedKey, _ in
-            if let property = properties[translatedKey],
-                property.type == .object,
-                let objectClassName = property.objectClassName,
-                let entityId = entity[Entity.EntityCodingKeys.entityId],
-                let dynamicObject = realm.dynamicObject(ofType: entityType, forPrimaryKey: entityId),
-                let nestedObject = dynamicObject[translatedKey] as? Object,
-                !(nestedObject is Entity)
-            {
-                cascadeDelete(
-                    realm: realm,
-                    entityType: objectClassName,
-                    entity: nestedObject
-                )
-            } else if let property = properties[translatedKey],
-                property.isArray,
-                let objectClassName = property.objectClassName,
-                let entityId = entity[Entity.EntityCodingKeys.entityId],
-                let dynamicObject = realm.dynamicObject(ofType: entityType, forPrimaryKey: entityId),
-                let nestedArray = dynamicObject[translatedKey] as? List<DynamicObject>
-            {
-                cascadeDelete(
-                    realm: realm,
-                    entityType: objectClassName,
-                    entities: nestedArray
-                )
-            }
-        }
-    }
-    
-    func save(entities: AnyRandomAccessCollection<JsonDictionary>, syncQuery: SyncQuery?) {
-        signpost(.begin, log: osLog, name: "Save Entities (JsonDictionary)")
-        defer {
-            signpost(.end, log: osLog, name: "Save Entities (JsonDictionary)")
-        }
-        log.debug("Saving \(entities.count) object(s)")
-        let propertyMapping = T.propertyMapping()
-        try! write { realm in
-            try entities.forEachAutoreleasepool { entity in
-                var translatedEntity = JsonDictionary()
-                try propertyMapping.forEachAutoreleasepool { (translatedKey, tuple) in
-                    let (key, transform) = tuple
-                    if let transform = transform,
-                        let value = transform.transformFromJSON(entity[key]) as? NSObject,
-                        let property = self.properties[translatedKey],
-                        !property.isArray,
-                        let objectClassName = property.objectClassName,
-                        let schema = realm.schema[objectClassName]
-                    {
-                        translatedEntity[translatedKey] = value.dictionaryWithValues(forKeys: schema.properties.map { $0.name })
-                    } else if self.needsTranslation,
-                        let array = entity[key] as? [Any],
-                        let property = self.properties[translatedKey],
-                        let objectClassName = property.objectClassName,
-                        typesNeedsTranslation.contains(objectClassName)
-                    {
-                        translatedEntity[translatedKey] = array.map {
-                            return ["value" : $0]
-                        }
-                    } else if let transform = transform {
-                        translatedEntity[translatedKey] = transform.transformFromJSON(entity[key])
-                    } else if let property = self.properties[translatedKey],
-                        !property.isArray,
-                        property.type == .object,
-                        let clazz = ObjCRuntime.typeForPropertyName(self.entityType, propertyName: translatedKey),
-                        let anyObjectClass = clazz as? (NSObject & JSONDecodable).Type,
-                        let json = entity[key] as? [String : Any]
-                    {
-                        var obj = anyObjectClass.init()
-                        try obj.refresh(from: json)
-                        translatedEntity[translatedKey] = obj
-                    } else {
-                        translatedEntity[translatedKey] = entity[key]
-                    }
-                }
-                self.cascadeDelete(
-                    realm: realm,
-                    entityType: self.entityTypeClassName,
-                    entity: entity,
-                    propertyMapping: propertyMapping
-                )
-                realm.dynamicCreate(self.entityTypeClassName, value: translatedEntity, update: .all)
-            }
-            self.saveQuery(syncQuery: syncQuery, realm: realm)
-        }
-    }
-    
-    func save(syncQuery: CacheType.SyncQuery) {
-        saveQuery(syncQuery: syncQuery, realm: self.newRealm)
-    }
-    
-    private func saveQuery(syncQuery: SyncQuery?, realm: Realm) {
-        guard let syncQuery = syncQuery else {
-            return
-        }
-        
-        let block: (String) -> Void = { entityTypeCollectionName in
-            let realmSyncQuery = _QueryCache()
-            realmSyncQuery.collectionName = entityTypeCollectionName
-            realmSyncQuery.query = syncQuery.query.predicate?.asString
-            realmSyncQuery.fields = syncQuery.query.fieldsAsString
-            realmSyncQuery.lastSync = syncQuery.lastSync
-            realmSyncQuery.generateKey()
-            realm.add(realmSyncQuery, update: .all)
-        }
-        if realm.isInWriteTransaction {
-            block(entityTypeCollectionName)
-        } else {
-            try! realm.write {
-                block(entityTypeCollectionName)
-            }
-        }
-    }
-    
 }
 
 extension AnyRandomAccessCollection where Element: NSObject, Element: Persistable {
@@ -968,218 +757,4 @@ extension AnyRandomAccessCollection where Element: NSObject, Element: Persistabl
         return self
     }
     
-}
-
-enum ObjectIdKind {
-    
-    case objectId(String)
-    case objectIds(AnyRandomAccessCollection<String>)
-    
-}
-
-extension ObjectIdKind {
-    
-    init?(_ objectId: String?) {
-        guard let objectId = objectId else {
-            return nil
-        }
-        self = .objectId(objectId)
-    }
-    
-    init?<RAC>(_ objectIds: RAC?) where RAC: RandomAccessCollection, RAC.Element == String {
-        guard let objectIds = objectIds else {
-            return nil
-        }
-        self = .objectIds(AnyRandomAccessCollection(objectIds))
-    }
-    
-}
-
-internal class RealmPendingOperation: Object, PendingOperation {
-    
-    @objc
-    dynamic var requestId: String = ""
-    
-    @objc
-    dynamic var date: Date = Date()
-    
-    @objc
-    dynamic var collectionName: String = ""
-    
-    @objc
-    dynamic var objectId: String?
-    
-    var objectIds: AnyRandomAccessCollection<String>?
-    var requestIds: AnyRandomAccessCollection<String>?
-    
-    @objc
-    dynamic var method: String = ""
-    
-    @objc
-    dynamic var url: String = ""
-    
-    @objc
-    dynamic var headers: Data = Data()
-    
-    @objc
-    dynamic var body: Data?
-    
-    convenience init<RequestIds>(
-        request: URLRequest,
-        collectionName: String,
-        objectIdKind: ObjectIdKind? = nil,
-        requestIds: RequestIds
-    ) where RequestIds: RandomAccessCollection, RequestIds.Element == String {
-        self.init(
-            request: request,
-            collectionName: collectionName,
-            objectIdKind: objectIdKind
-        )
-        self.requestIds = AnyRandomAccessCollection(requestIds)
-    }
-    
-    convenience init(
-        request: URLRequest,
-        collectionName: String,
-        objectIdKind: ObjectIdKind? = nil
-    ) {
-        self.init()
-        
-        requestId = request.value(forHTTPHeaderField: KinveyHeaderField.requestId)!
-        self.collectionName = collectionName
-        if let objectIdKind = objectIdKind {
-            switch objectIdKind {
-            case .objectId(let objectId):
-                self.objectId = objectId
-            case .objectIds(let objectIds):
-                self.objectIds = objectIds
-            }
-        }
-        method = request.httpMethod ?? "GET"
-        url = request.url!.absoluteString
-        headers = try! JSONSerialization.data(withJSONObject: request.allHTTPHeaderFields!)
-        body = request.httpBody
-    }
-    
-    func buildRequest() -> URLRequest {
-        var request = URLRequest(url: URL(string: url)!)
-        request.httpMethod = method
-        request.allHTTPHeaderFields = try? JSONSerialization.jsonObject(with: headers) as? [String : String]
-        if let body = body {
-            request.httpBody = body
-        }
-        return request
-    }
-    
-    override class func primaryKey() -> String? {
-        return "requestId"
-    }
-    
-}
-
-class RealmPendingOperationReference: PendingOperation {
-    
-    let realmConfig: Realm.Configuration
-    let requestId: String
-    
-    init(_ realmPendingOperation: RealmPendingOperation) {
-        realmConfig = realmPendingOperation.realm!.configuration
-        requestId = realmPendingOperation.requestId
-    }
-    
-    lazy var realmPendingOperation: RealmPendingOperation = {
-        let realm = try! Realm(configuration: realmConfig)
-        return realm.object(ofType: RealmPendingOperation.self, forPrimaryKey: requestId)!
-    }()
-    
-    var collectionName: String {
-        return realmPendingOperation.collectionName
-    }
-    
-    var objectId: String? {
-        return realmPendingOperation.objectId
-    }
-    
-    var objectIds: AnyRandomAccessCollection<String>? {
-        return realmPendingOperation.objectIds
-    }
-    
-    var requestIds: AnyRandomAccessCollection<String>? {
-        return realmPendingOperation.requestIds
-    }
-    
-    func buildRequest() -> URLRequest {
-        return realmPendingOperation.buildRequest()
-    }
-    
-}
-
-internal class _QueryCache: Object {
-    
-    @objc
-    dynamic var key: String?
-    
-    internal func generateKey() {
-        key = "\(collectionName ?? "nil")|\(query ?? "nil")"
-    }
-    
-    @objc
-    dynamic var collectionName: String?
-    
-    @objc
-    dynamic var query: String?
-    
-    @objc
-    dynamic var fields: String?
-    
-    @objc
-    dynamic var lastSync: Date?
-    
-    @objc
-    override class func primaryKey() -> String? {
-        return "key"
-    }
-    
-}
-
-class RealmCascadeDeleteExecutor: CascadeDeleteExecutor {
-    
-    let realm: Realm
-    
-    init(realm: Realm) {
-        self.realm = realm
-    }
-    
-    func cascadeDelete<Value>(_ object: Value?) throws where Value : Object {
-        if let object = object {
-            if let cascadeDelete = object as? CascadeDeletable {
-                try cascadeDelete.cascadeDelete(executor: self)
-            }
-            deleteAclAndMetadata(realm: realm, object: object)
-            realm.delete(object)
-        }
-    }
-    
-    func cascadeDelete<Value>(_ list: List<Value>) throws where Value : Object {
-        for object in list {
-            if let cascadeDelete = object as? CascadeDeletable {
-                try cascadeDelete.cascadeDelete(executor: self)
-            }
-            deleteAclAndMetadata(realm: realm, object: object)
-        }
-        realm.delete(list)
-    }
-    
-}
-
-fileprivate func deleteAclAndMetadata<T>(realm: Realm, object: T) where T: Object {
-    guard let entity = object as? Entity else {
-        return
-    }
-    if let acl = entity.acl {
-        realm.delete(acl)
-    }
-    if let metadata = entity.metadata {
-        realm.delete(metadata)
-    }
 }
